requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymysql==1.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
openai==1.3.7
scikit-learn==1.3.2
numpy>=1.26.0
pandas>=2.1.0
Pillow==10.1.0

# mysqlclient는 Windows에서 설치가 어려울 수 있으므로 pymysql 사용
# mysqlclient==2.2.0  # 필요시 주석 해제

